fix: add golden tests for registry module
