+++ .github/workflows/go.yml
@@ -36,6 +36,7 @@ jobs:
         env:
           INFRACOST_API_KEY: "00000000000000000000000000000000"
           INFRACOST_LOG_LEVEL: info
+          INFRACOST_TERRAFORM_CLOUD_TOKEN: ${{ secrets.TERRAFORM_CLOUD_TOKEN }}
 
       - name: Test example works ok from CLI
         run: |
