+++ cmd/infracost/breakdown_test.go
@@ -363,3 +363,16 @@ func TestBreakdownInitFlagsError(t *testing.T) {
 		nil,
 	)
 }
+
+func TestBreakdownWithPrivateTerraformRegistryModule(t *testing.T) {
+	GoldenFileCommandTest(
+		t,
+		testutil.CalcGoldenFileTestdataDirName(),
+		[]string{
+			"breakdown",
+			"--path",
+			path.Join("./testdata", testutil.CalcGoldenFileTestdataDirName()),
+		},
+		nil,
+	)
+}
