+++ cmd/infracost/testdata/breakdown_with_private_terraform_registry_module/main.tf
@@ -1,0 +1,27 @@
+provider "aws" {
+  region                      = "us-east-1" # <<<<< Try changing this to eu-west-1 to compare the costs
+  skip_credentials_validation = true
+  skip_requesting_account_id  = true
+  access_key                  = "mock_access_key"
+  secret_key                  = "mock_secret_key"
+}
+
+module "ec2_cluster" {
+  source  = "app.terraform.io/infracost/ec2-instance/aws"
+  version = "~> 2.0"
+
+  name           = "my-cluster"
+  instance_count = 5
+
+  ami                    = "ami-ebd02392"
+  instance_type          = "t2.micro"
+  key_name               = "user1"
+  monitoring             = true
+  vpc_security_group_ids = ["sg-12345678"]
+  subnet_id              = "subnet-eddcdzz4"
+
+  tags = {
+    Terraform   = "true"
+    Environment = "dev"
+  }
+}
