+++ examples/terraform/main.tf
@@ -6,24 +6,28 @@ provider "aws" {
   secret_key                  = "mock_secret_key"
 }
 
-module "ec2_cluster" {
-  source  = "app.terraform.io/infracost/ec2-instance/aws"
-  version = "~> 2.0"
+resource "aws_instance" "web_app" {
+  ami           = "ami-674cbc1e"
+  instance_type = "m5.4xlarge" # <<<<< Try changing this to m5.8xlarge to compare the costs
 
-  name           = "my-cluster"
-  instance_count = 5
-
-  ami                    = "ami-ebd02392"
-  instance_type          = "t2.micro"
-  key_name               = "user1"
-  monitoring             = true
-  vpc_security_group_ids = ["sg-12345678"]
-  subnet_id              = "subnet-eddcdzz4"
-
-  tags = {
-    Terraform   = "true"
-    Environment = "dev"
+  root_block_device {
+    volume_size = 50
   }
+
+  ebs_block_device {
+    device_name = "my_data"
+    volume_type = "io1" # <<<<< Try changing this to gp2 to compare costs
+    volume_size = 1000
+    iops        = 800
+  }
+}
+
+resource "aws_lambda_function" "hello_world" {
+  function_name = "hello_world"
+  role          = "arn:aws:lambda:us-east-1:account-id:resource-id"
+  handler       = "exports.test"
+  runtime       = "nodejs12.x"
+  memory_size   = 1024 # <<<<< Try changing this to 512 to compare costs
 }
 
 output "aws_instance_type" {
